fix: D3D11VA compile errors on Windows
- render_d3d11_resident slice params: u32 -> usize (matches
  StereoFrame::D3d11Resident field types and stage_frame signature)
- as_hal::<Dx12>() wrapped in unsafe block (Rust 2024 edition
  requires explicit unsafe inside unsafe fn)

+++ crates/reco-core/src/session/frame_processing.rs
@@ -402,9 +402,9 @@ impl StitchSession {
     fn render_d3d11_resident(
         &mut self,
         left_texture: *mut std::ffi::c_void,
-        left_slice: u32,
+        left_slice: usize,
         right_texture: *mut std::ffi::c_void,
-        right_slice: u32,
+        right_slice: usize,
         yaw: f32,
         pitch: f32,
     ) -> Result<(), SessionError> {
